refactor: Revalidate data cache every 3600s and fetch data for Products
- Configure the homepage to revalidate the data cache every 3600 seconds.
- Change the Products section to server components.
- Implement interactivity to display products by category on the ProductsCategories component.
