+++ app/page.tsx
@@ -4,6 +4,8 @@ import Products from "@/views/Products";
 import AboutUs from "@/views/AboutUs";
 import Testimoni from "@/views/Testimoni";
 
+export const revalidate = 3600;
+
 export default function Home() {
   return (
     <>
